+++ src/app/Pages/paid-listing/paid-listing.css
@@ -1,132 +1,130 @@
 .custom-section {
-    /* background: #f9f9f9; */
-    /* padding: 30px 0; */
+  /* background: #f9f9f9; */
+  /* padding: 30px 0; */
 }
 
 .paid-listing-image {
-    height: 50px;
-    width: 50px;
-    object-fit: cover;
-    border-radius: 50%;
+  height: 50px;
+  width: 50px;
+  object-fit: cover;
+  border-radius: 50%;
 }
 
 .listing-image img {
-    height: 120px;
-    width: 100%;
-    object-fit: cover;
-    border-radius: 10px;
+  height: 120px;
+  width: 100%;
+  object-fit: cover;
+  border-radius: 10px;
 }
 
 /* paid-listing.css */
 
 .custom-title {
-    text-align: center;
-    color: var(--black);
-    margin-bottom: 30px;
-    font-size: 24px;
+  text-align: center;
+  color: var(--black);
+  margin-bottom: 30px;
+  font-size: 24px;
 }
 
 .custom-col {
-    display: flex;
-    justify-content: space-between;
-    width: 100%;
-    align-items: center;
-    box-shadow: 0px 0px 2px 0px lightgray;
-    padding: 15px;
-    border-radius: 20px;
-    background: white;
-    margin-bottom: 1rem;
+  display: flex;
+  justify-content: space-between;
+  width: 100%;
+  align-items: center;
+  box-shadow: 0 1px lightgray, 0 -1px lightgray;
+  padding: 15px;
+  border-radius: 20px;
+  background: white;
+  margin-bottom: 1rem;
 }
 
 .verified-text {
-    background-color: green;
-    color: white;
-    padding: 3px 6px;
-    border-radius: 5px;
-    font-size: 9px;
-    font-weight: 400;
-    margin: 0;
-    line-height: 0;
-    width: fit-content;
+  background-color: green;
+  color: white;
+  padding: 3px 6px;
+  border-radius: 5px;
+  font-size: 9px;
+  font-weight: 400;
+  margin: 0;
+  line-height: 0;
+  width: fit-content;
 }
 
 .custom-row {
-    margin-bottom: 1rem;
+  margin-bottom: 1rem;
 }
-
 
 .listing-content {
-    display: flex;
-    gap: 10px;
-    align-items: center;
-    margin-bottom: 0px;
+  display: flex;
+  gap: 10px;
+  align-items: center;
+  margin-bottom: 0px;
 }
 
-
 .listing-content a {
-    text-decoration: none;
+  text-decoration: none;
 }
 
 .listing-content a:nth-child(1) {
-    color: var(--black);
-    font-size: 16px;
-    font-weight: 600;
-    margin: 0;
+  color: var(--black);
+  font-size: 16px;
+  font-weight: 600;
+  margin: 0;
 }
 
 .listing-content a:nth-child(2) {
-    font-size: 12px;
-    margin: 0;
-    color: var(--black);
+  font-size: 12px;
+  margin: 0;
+  color: var(--black);
 }
 
 .listing-title {
-    font-size: 18px;
-    font-weight: 400;
-    color: var(--black);
-    text-decoration: none;
+  font-size: 18px;
+  font-weight: 400;
+  color: var(--black);
+  text-decoration: none;
 }
 
 .listing-title:hover {
-    text-decoration: underline;
+  text-decoration: underline;
 }
 
 .listing-description,
 .listing-rating {
-    max-width: 90%;
-    margin: 0;
-    color: #555;
-    font-size: 12px;
+  max-width: 90%;
+  margin: 0;
+  color: #555;
+  font-size: 12px;
 }
 
 .listing-rating i {
-    color: green;
+  color: green;
 }
 
-@media only screen and (max-width:768px) {
-    .custom-col {
-        width: 100%;
-        box-shadow: 0px 0px 6px -3px lightgray;
-    }
+@media only screen and (max-width: 768px) {
+  .custom-col {
+    width: 100%;
+    box-shadow: 0px 0px 6px -3px lightgray;
+  }
 
-    .listing-image img {
-        height: 80px;
-        width: 100%;
-        object-fit: cover;
-        border-radius: 5px;
-    }
+  .listing-image img {
+    height: 80px;
+    width: 100%;
+    object-fit: cover;
+    border-radius: 5px;
+  }
 
-    .listing-description,
-    .listing-rating {
-        max-width: 100%;
-        font-size: 10px;
-    }
+  .listing-description,
+  .listing-rating {
+    max-width: 100%;
+    font-size: 10px;
+  }
 
-    .listing-title {
-        color: var(--blue);
-        font-size: 16px;
-    }
-    .custom-section {
-        padding: 0;
-    }
+  .listing-title {
+    color: var(--blue);
+    font-size: 16px;
+  }
+  .custom-section {
+    padding: 0;
+  }
 }
