+++ src/app/globals.css
@@ -46,7 +46,7 @@ body {
   padding: 8px 15px;
   border-radius: 30px;
   color: var(--black);
-  transition: 0.5s;
+  /* transition: 0.5s; */
   text-decoration: none;
 }
